requirements.txt
-e .
python-dotenv
fastapi
uvicorn
google-generativeai==0.5.2
google-cloud-texttospeech
pydub